fix: Resolved minor issues
