+++ src/controllers/openid4vc/holder/holder.Controller.ts
@@ -5,7 +5,6 @@ import { injectable } from 'tsyringe'
 import { SCOPES } from '../../../enums/enum'
 import {
   AuthorizeRequestCredentialOffer,
-  CredentialType,
   DeleteCredentialBody,
   RequestCredentialBody,
   ResolveCredentialOfferBody,
